Add authentication

+++ packages/server/src/routes/jobApplication.route.js
@@ -1,14 +1,21 @@
-const express = require('express');
-const router = express.Router();
+const router = require('express').Router();
 const JobApplicationModel = require('../models/jobApplication.model');
 
 router.post('/', (req, res) => {
+  if (!req.user) {
+    return res.status(401).json({
+      success: false,
+      message: 'Unauthorized',
+    });
+  }
+
   const body = req.body;
   const info = req.user;
-  console.log('INFO USER', info);
-  console.log('INFO BODY', body);
 
-  JobApplicationModel.create(body)
+  JobApplicationModel.create({
+    ...body,
+    ownerId: req.user._id,
+  })
     .then((application) => {
       res.json(application);
     })
@@ -18,6 +25,13 @@ router.post('/', (req, res) => {
 });
 
 router.put('/:id', (req, res) => {
+  if (!req.user) {
+    return res.status(401).json({
+      success: false,
+      message: 'Unauthorized',
+    });
+  }
+
   const id = req.params.id;
   const body = req.body;
   JobApplicationModel.findByIdAndUpdate(id, { $set: body })
@@ -30,6 +44,13 @@ router.put('/:id', (req, res) => {
 });
 
 router.delete('/:id', async function (req, res) {
+  if (!req.user) {
+    return res.status(401).json({
+      success: false,
+      message: 'Unauthorized',
+    });
+  }
+
   const id = req.params.id;
   try {
     await JobApplicationModel.deleteOne({ _id: id });
